slight update

+++ ifstatement/dataset/data_process.py
@@ -7,7 +7,9 @@ from tqdm import tqdm
 import random
 import datetime
 import pytz
+import pickle
 
+# from repo_names import REPO_NAMES
 
 def extract_and_rename_python_files(repo_url, destination_folder, index, index_dic, progress_bar=None):
     tmp_folder = './tmp'
@@ -27,7 +29,7 @@ def extract_and_rename_python_files(repo_url, destination_folder, index, index_d
             ["git", "clone", repo_url, repo_path],
             stdout=subprocess.DEVNULL,
             stderr=subprocess.DEVNULL,
-            timeout=10
+            timeout=20
         )
     except subprocess.TimeoutExpired:
         print(f"Error: Cloning the repository {repo_url} took too long (more than 10 seconds).")
@@ -44,10 +46,12 @@ def extract_and_rename_python_files(repo_url, destination_folder, index, index_d
     if not os.path.exists(destination_folder):
         os.makedirs(destination_folder)
 
-    print(f"python_files: {python_files}")
+    # print(f"python_files: {python_files}")
 
     for python_file in python_files:
-        new_filename = f"{index:08d}.py"
+        if " " in python_file:
+            continue
+        new_filename = f"{index:08d}.txt"
         destination_path = os.path.join(destination_folder, new_filename)
         # shutil.copy(python_file, destination_path)
         try:
@@ -63,10 +67,10 @@ def extract_and_rename_python_files(repo_url, destination_folder, index, index_d
                 print(f"index: {index:08d} func: {num_of_func:03d} filename: {python_file[6:]}")
             else:
                 progress_bar.set_description(f"index: {index:08d} func: {num_of_func:03d} filename: {python_file[6:]}")
-                print(f"index: {index:08d} func: {num_of_func:03d} filename: {python_file[6:]}")
+                # print(f"index: {index:08d} func: {num_of_func:03d} filename: {python_file[6:]}")
             index += 1
-
-    shutil.rmtree(repo_path)
+    if os.path.exists(repo_path):
+        shutil.rmtree(repo_path)
     return index
 
 
@@ -76,7 +80,7 @@ def process_python_file(python_file, destination_path):
         content = f.read()
 
     # Remove single-line comments
-    content = re.sub(r'#.*?(\n|$)', '', content)
+    # content = re.sub(r'#.*?(\n|$)', '', content)
     # print("=" * 200)
     # print(content)
 
@@ -92,11 +96,15 @@ def process_python_file(python_file, destination_path):
         func_start = func_match.start()
         func_end = find_func_end(content, func_start)
         function_code = content[func_start:func_end]
-        print("=" * 200)
-        print(function_code)
-        if re.search(r'^\n[ \t]*if', function_code, re.MULTILINE):  # if re.match(r'^\n[ \t]*if', function_code):
+        if len(function_code) > 2000 or len(function_code) < 500:
+            continue
+        if not has_one_def_keyword(function_code):
+            continue
+        # print("=" * 200)
+        # print(function_code)
+        if re.search(r'\n[ \t]*if', function_code, re.MULTILINE):  # if re.match(r'^\n[ \t]*if', function_code):
             num_of_func += 1
-            new_destination_path = destination_path.replace(".py", f"_{num_of_func:04d}.py")
+            new_destination_path = destination_path.replace(".txt", f"_{num_of_func:04d}.txt")
             # functions_content += function_code + "\n\n"  # Add some spacing between functions
 
             # Write the extracted functions to the destination file
@@ -105,6 +113,15 @@ def process_python_file(python_file, destination_path):
 
     return num_of_func
 
+
+def has_one_def_keyword(function_code):
+    # Regex pattern to match "def" at the beginning of a line or following whitespace
+    # and followed by a space to ensure it's not part of another word.
+    pattern = r'(^|\s)def\s'
+    # Find all matches of the pattern in the string
+    matches = re.findall(pattern, function_code)
+    # Return True if there's only one match, otherwise False
+    return len(matches) == 1
 
 def find_func_end(content, func_start):
     """
@@ -137,7 +154,7 @@ def find_func_end(content, func_start):
 
 
 def build_dataset_files(raw_repo_names, start, end):
-    repo_name_list = raw_repo_names.split()
+    repo_name_list = raw_repo_names  # .split()
     repo_name_list = repo_name_list[start:end]
     index_save_path = "index.json"
 
@@ -156,14 +173,28 @@ def build_dataset_files(raw_repo_names, start, end):
 
     with tqdm(total=len(repo_name_list)) as progress_bar:
         # for one_repo_name in tqdm(repo_name_list):
-        for one_repo_name in repo_name_list:
+        for idx, one_repo_name in enumerate(repo_name_list):
             if one_repo_name in index_dic["repo_list"]:
                 print(f"skip repo {one_repo_name}")
+                progress_bar.update(1)
+                continue
             else:
                 repo_url = f"https://github.com/{one_repo_name}"
                 index = extract_and_rename_python_files(repo_url, "data/", index, index_dic, progress_bar)
             progress_bar.update(1)
             index_dic["repo_list"].append(one_repo_name)
+            if idx % 20 == 0:
+                index_dic["index"] = index
+
+                # sum_class = sum([index_dic["class"][str(i)] for i in range(index) if str(i) in index_dic["class"]])
+                sum_func = sum([index_dic["func"][str(i)] for i in range(index) if str(i) in index_dic["func"]])
+
+                # index_dic["sum_class"] = sum_class
+                index_dic["sum_func"] = sum_func
+                with open(index_save_path, "w") as f:
+                    json.dump(index_dic, f, indent=4)
+
+                print(f"idx = {idx} Index file saved to {index_save_path}")
     index_dic["index"] = index
 
     # sum_class = sum([index_dic["class"][str(i)] for i in range(index) if str(i) in index_dic["class"]])
@@ -174,11 +205,13 @@ def build_dataset_files(raw_repo_names, start, end):
 
     print(f"sum_func: {sum_func}")
 
-
     with open(index_save_path, "w") as f:
         json.dump(index_dic, f, indent=4)
 
     print(f"Index file saved to {index_save_path}")
 
+
 if __name__ == "__main__":
-    build_dataset_files("0215Arthur/HG-GNN", 0, 1)
+    with open("ifstatement/dataset/repo_names.pkl", "rb") as f:
+        repo_names = pickle.load(f)
+    build_dataset_files(repo_names, 8500, 30000)
